backend/requirements.txt
flask==2.2.3
werkzeug==2.2.3
flask-cors==3.0.10
openai==0.27.0
python-dotenv==1.0.0
bcrypt==4.0.1
